ioctl cmd arg is u_long, strategy returns void.
SCCS-vsn: sys/hp300/dev/dcm.c 8.5
SCCS-vsn: sys/hp300/dev/sd.c 8.6
SCCS-vsn: sys/hp300/dev/ppi.c 8.2
SCCS-vsn: sys/hp300/dev/rd.c 8.3
SCCS-vsn: sys/hp300/dev/ct.c 8.3
SCCS-vsn: sys/hp300/dev/st.c 8.5

+++ usr/src/sys/hp300/dev/dcm.c
@@ -11,7 +11,7 @@
  *
  * from Utah: $Hdr: dcm.c 1.29 92/01/21$
  *
- *	@(#)dcm.c	8.4 (Berkeley) %G%
+ *	@(#)dcm.c	8.5 (Berkeley) %G%
  */
 
 /*
@@ -699,7 +699,7 @@ dcmmint(unit, mcnd, dcm)
 
 dcmioctl(dev, cmd, data, flag, p)
 	dev_t dev;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	int flag;
 	struct proc *p;
