+++ usr/src/sys/hp300/dev/rd.c
@@ -11,7 +11,7 @@
  *
  * from: Utah $Hdr: rd.c 1.44 92/12/26$
  *
- *	@(#)rd.c	8.2 (Berkeley) %G%
+ *	@(#)rd.c	8.3 (Berkeley) %G%
  */
 
 /*
@@ -906,7 +906,7 @@ rdwrite(dev, uio, flags)
 int
 rdioctl(dev, cmd, data, flag, p)
 	dev_t dev;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	int flag;
 	struct proc *p;
